+++ content/themes/make/inc/customizer/bootstrap.php
@@ -15,7 +15,6 @@ function ttfmake_customizer_init() {
 	$path = get_template_directory() . '/inc/customizer/';
 
 	// Always load
-	require_once( $path . 'compatibility.php' );
 	require_once( $path . 'controls.php' );
 	require_once( $path . 'helpers.php' );
 	require_once( $path . 'helpers-css.php' );
@@ -271,12 +270,26 @@ function ttfmake_customizer_add_section_options( $section, $args, $initial_prior
 		// Add setting
 		if ( isset( $option['setting'] ) ) {
 			$defaults = array(
-				'default' => ttfmake_get_default( $setting_id ),
-				'type'    => 'theme_mod',
+				'type'                 => 'theme_mod',
+				'capability'           => 'edit_theme_options',
+				'theme_supports'       => '',
+				'default'              => ttfmake_get_default( $setting_id ),
+				'transport'            => 'refresh',
+				'sanitize_callback'    => '',
+				'sanitize_js_callback' => '',
 			);
 			$setting = wp_parse_args( $option['setting'], $defaults );
 
-			$wp_customize->add_setting( $setting_id, $setting );
+			// Add the setting arguments inline so Theme Check can verify the presence of sanitize_callback
+			$wp_customize->add_setting( $setting_id, array(
+				'type'                 => $setting['type'],
+				'capability'           => $setting['capability'],
+				'theme_supports'       => $setting['theme_supports'],
+				'default'              => $setting['default'],
+				'transport'            => $setting['transport'],
+				'sanitize_callback'    => $setting['sanitize_callback'],
+				'sanitize_js_callback' => $setting['sanitize_js_callback'],
+			) );
 		}
 
 		// Add control
@@ -298,10 +311,16 @@ function ttfmake_customizer_add_section_options( $section, $args, $initial_prior
 			// Check for a specialized control class
 			if ( isset( $control['control_type'] ) ) {
 				$class = $control['control_type'];
-				unset( $control['control_type'] );
-				$wp_customize->add_control(
-					new $class( $wp_customize, $control_id, $control )
-				);
+
+				if ( class_exists( $class ) ) {
+					unset( $control['control_type'] );
+
+					// Dynamically generate a new class instance
+					$reflection = new ReflectionClass( $class );
+					$class_instance = $reflection->newInstanceArgs( array( $wp_customize, $control_id, $control ) );
+
+					$wp_customize->add_control( $class_instance );
+				}
 			} else {
 				$wp_customize->add_control( $control_id, $control );
 			}